refactor: consolidate UP handling
Before these changes, we were performing Unit Primitive handling in several
places in the code. This led to some inconsistencies in the code, e.g.:

- A UP being listed in the prompt .yml definition but never checked,
or a different one being checked
- Internal events being tracked on some UP checks but not others

This change consolidates checking if the user has access to all UPs, and
registering the proper internal events, in the `BasePromptRegistry.get_on_behalf`
method. We change all runtime code that was calling `get` to `get_on_behalf`.