chore: update default prompts example
I added this to illustrate the json format of `AIGW_DEFAULT_PROMPTS`,
but if you just `cp example.env .env` you end up with a random
override for no good reason. Plus we removed the vertex prompt
referenced in the current example in
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/merge_requests/1970,
so even worse now you'd get a broken setup. So I update the example
with a still valid prompt, but leave it commented out.