feat: agentic duo chat goal parser module implementation
This merge request adds in the logic and test cases for the
SlashCommandsGoalParser module. This module is responsible for taking in
the user's goal and determining which slash command is being requested.

It will then separate the slash command from any additional text that is
added after the slash command.

There is also some basic error handling for common typo situations.

Fix linting error and removes import statment

This commit simply removes an unused import statement and adds in a skip
for the linting check of the test file name.

Fix import statements on prompt expander test

This commit fixes the import statements for the SlashCommandsGoalParser
module that was within the test file for the SlashCommandsPromptExpander
test suite.