chore: fix test_workflow_run mock any number of times
Right now this test is flaky. See this
[Slack thread (internal)](https://gitlab.slack.com/archives/C08JFG280BX/p1747801510383489)

It's not totally obvious why this seems to run a different number of
times in CI vs. locally but we think it has to do with some env vars
that control the behaviour of handover.

This fixes it for now by just allowing it to be called any number of
times but we probably want to follow up and make this test consistent in
different environments because different behaviour in CI vs. locally
makes debugging failed CI very tricky.