feat: update to slash commands prompt expander
This commit updates the SlashCommandsPromptExpander class to include the
changes from updating the SlashCommandsGoalParser module to no longer
have a class and instead have a function called parse that is now used.

This also updates the test cases to work with the new parse method and
cleans up the test cases.