chore(lint): install all dependencies when linting
In CI we were running `poetry install --only lint`, whereas locally
developers were running `poetry install`, which meant that some
mypy errors did not show up on CI because it couldn't look into some
of the dependencies. We change the lint install to `--with lint`,
and change the CI config to make it depend on the `install` job.