fix: remove system prompt from definition
This commit removes the system prompt being loaded into the
SlashCommandsDefinition module. The system prompt will not be updated
with each slash command. Instead we will be updating the goal in a way
that accomplishes the same thing that the system prompts were doing in
the standard Duo Chat slash commands.