chore(lint): re-enable various linters
To keep this commit short, we up the limits of many linters to match
the current maximums, and at least make sure that future code doesn't
push these metrics further.

We also install eval packages in the linting stage so the linter can
look at the libraries for that code. In a similar vein, we allow the
`dependency_injector` C extension to be loaded by the linter.

Lastly, we ignore protobuf autogenerated packages (see comment in
pyproject.toml)