chore(tests): avoid dependency on ENV in DWS tests
The current tests would fail if DUO_WORKFLOW__VERTEX_PROJECT_ID=""
and ANTHROPIC_API_KEY=="". Tests should be independent of the current
environment, so we mock `create_chat_model` in the affected tests.
We missed this because for GitLab developers and in the main repo
these variables are set, but we were seeing failures in the community
fork (for example
https://gitlab.com/gitlab-community/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/jobs/10721335360).