chore: collapse more gRPC file types by default
At
https://docs.gitlab.com/user/project/merge_requests/changes/#collapse-generated-files
you can see that GitLab has a feature for collapsing files automatically
in the diff view when reviewing MRs. It's useful to hide generated files
from diffs as they tend to be large and distract from reviewing the
important code.

We were already handling most of our generated gRPC files in here but we
missed a couple.