refactor: revert "add consistent expanded logging of model metadata for endpoints"
In https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/merge_requests/2628
we tried to tackle the lack of model-related labels in logging entries by using the ModelMetadata
class to populate the request context. But this is not its purpose, and it can lead to building
objects that are not valid model metadata (for example, we used "default" as a model provider in
some cases, which is not a valid provider). It also ignores the fact that a request may be served
by multiple LLMs. We should instead rely on adding the relevant fields to the log entries that are
in the context of a given LLM call