feat: use different API for get-file tool
The current tool implementation doesn't work if an API call
is not performed to Gitlab directly but rather proxied via
Workhorse via sending runHTTPRequest gRPC message.

Using a different endpoint fixes the issue. The other endpoint
returns encoded base64 content though, so we need to decode it
to make it work. The change should be seemless to the user