feat(proxy): extend the list of accepted unit primitives
Now the proxy endpoint accepts ai_gateway_model_provider_proxy

It's implemented in AI Gateway, because eventually we want to
restrict this endpoint to this unit primitive only. Currently,
this endpoint is used by older self-managed instances, so we
cannot just restrict the list of unit primitives due to the
backward compatibility promise.