chore: add logging to indicate when we are sending checkpoint notifies
In an attempt to better debug
https://gitlab.com/groups/gitlab-org/-/epics/19253#note_2743690530, which
appears to possibly be a race condition between `NewCheckpoint` and
making HTTP requests, this MR introduces some more logging around the
checkpoint behaviour.

This is line with the other outgoing actions in
duo_workflow_service/executor/action.py
where we log every other kind of outgoing gRPC action.

This logging may be quite a bit noisier as we send a lot of checkpoints
when streaming long LLM responses but we it seems important enough for
now that we should log it to help us debug these issues. We can consider
removing it later if it is too noisy and if it is no longer needed for
debugging issues.