chore: add yet more logging to debug agentic chat blocking
This expands on the logs we added recently to see conclusively that the
Duo Workflow Service is indeed blocked waiting for a gRPC message.

Our existing logs show that workhorse thinks it is sending the message
and the last logs from Duo Workflow Service do indicate it's blocked at
this point but we want to be extra sure that we're not in one of these
branches of logic that ignores the messages.