fix(workflow): increase timeouts for workflow prompts
Currently, we enforce 30 seconds timeout for Anthropic Python SDK client.

There's some specifics in Anthropic API server behavior:
it stops streaming at some point for large input/outputs.

However, it sends a ping message every 30 seconds. As a result,
we have a race condition and sometimes the ping message is not processed
within the 30 seconds timeout window. Let's increase the timeout
window to 60 seconds. In this case, we'll have enough time to process
the ping message if it's send by the Anthropic server.

This commit removes the timeout overrides and uses the default value
for Anthropic client: 60 seconds.