chore: always send requestID with outgoing actions
At present we are not sending the requestID at all. We also weren't
using this for anything before. As part of
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/merge_requests/3268
we are going to start relying on this requestID to correlate responses
with requests out over gRPC.

That MR is also introducing the requestID but it may take longer to get
merged. In the meantime this requestID is still very useful for logging.
It's also being logged by workhorse since
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/208056 .

As such I wanted to make this small change to include the requestID
always so that we can correlate logs.

Unfortunately this change does not implement this in the exact same way
as
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/merge_requests/3268
because that MR is introducing it in a new file and it doesn't make
sense to include that new file in this MR.