chore: replace StrOutputParser with .text call
StrOutputParser.invoke seems to contribute a lot to our CPU
usage because it does a lot of additional steps related to
callbacks calling.

We don't need this additional logic here, we just need to extract
content from the message.

Let's replace it with .text which is much simpler and does
exactly what we need