chore: configure DWS runway with less CPU and memory and
Related to
https://gitlab.com/gitlab-org/gitlab/-/issues/578167#note_2845880895
we found that we may see better autoscaling characteristics with 1 vCPU
instead of 2. That's because runway scales up instances when an instance
reaches 60% CPU. But asyncio python applications run single threaded so
this won't happen if we have 2 vCPU available.

Also I reduced memory to 4GB and instance request concurrency to 45
as they matched our better benchmarking results.