chore: correctly track CancelledError in metrics
As discussed in #1591
we were seeing an issue where our prometheus metrics were never
capturing the CANCELLED gRPC status code. This was due to the fact that
we are catching `Exception` but `asyncio.CancelledError` inherits from
`BaseException` and not `Exception` as per the docs at
https://docs.python.org/3/library/asyncio-exceptions.html#asyncio.CancelledError
. As such all of these exceptions were being raised out of our metrics
instrumentation and never getting incremented in prometheus.

Once we deploy this change we should start to see a lot of `CANCELLED`
grpc statuses in our metrics and the overall error rates will increase
as these were never being captured before.